Split KeyPairAlgorithm, make non-CLI part public
